show task in dashbord of manager and user

+++ src/main/java/org/example/devsync1/services/UserService.java
@@ -1,10 +1,12 @@
 package org.example.devsync1.services;
 
 import org.example.devsync1.entities.User;
+import org.example.devsync1.enums.Role;
 import org.example.devsync1.repositories.implementations.UserRepository;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 public class UserService {
     UserRepository userRepository;
@@ -18,16 +20,28 @@ public class UserService {
         return userRepository.findById(id);
     }
     public List<User> findAll() {
-        return userRepository.findAll();
+        return userRepository.findAll()
+                .stream()
+                .filter(user -> user.getRole().equals(Role.USER))
+                .collect(Collectors.toList());
     }
     public User findByEmail(String email) {
         return userRepository.findByEmail(email);
     }
     public void update(User user) {
-        userRepository.update(user);
+        User userfound = this.findById(user.getId());
+        if (userfound != null) {
+            userRepository.delete(user);
+        }else {
+            System.out.println("User not found");
+        }
     }
-
     public void delete(User user) {
-        userRepository.delete(user);
+        User userfound = this.findById(user.getId());
+        if (userfound != null) {
+            userRepository.delete(user);
+        }else {
+            System.out.println("User not found");
+        }
     }
 }
